requirements.txt
streamlit
python-dotenv
google-generativeai
pypdf

